requirements/local.txt
# Local development dependencies go here
-r test.txt
django-debug-toolbar==1.2.1
Sphinx==1.2.2
mock==1.0.1
